+++ types/interfaces.d.ts
@@ -10,12 +10,6 @@ declare module 'fastify' {
             convertValidationErrors?: boolean;
         };
     }
-    interface FastifyReply {
-        originalResponse?: {
-            statusCode: number;
-            payload: any;
-        };
-    }
 }
 export interface GenericObject {
     [key: string]: any;