+++ types/properties.d.ts
@@ -1,3 +1,0 @@
-import { GenericObject } from './interfaces';
-export declare function addAdditionalProperties(target: GenericObject, source: GenericObject): void;
-export declare function serializeError(error: Error): GenericObject;